06.14.15
Fixed a glitch where the looted items were creating a new item list and
deleting the previous items each time.

After testing all classes seem to be operational (for now) will finish
story. Also will add a map of each equipped item, so that we can check
if the item is already equipped, and not allow another.